Separating solver metrics calculation to metrics_calculator
